refactor: merge PartiesUI into Parties
